pyproject.toml
[project]
name = "l3april2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "certifi>=2026.4.22",
    "ipykernel>=7.2.0",
    "langchain>=1.2.16",
    "langchain-aws>=1.4.6",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.2.1",
    "langchain-tavily>=0.2.18",
    "pip-system-certs>=5.3",
    "python-dotenv>=1.2.2",
]
